requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
pydantic-settings>=2.2
pydantic[email]>=2.0
jinja2>=3.1
python-multipart>=0.0.9
httpx>=0.27
itsdangerous>=2.1
